Fixed std::complex*int bug in SpecBit

+++ SpecBit/src/SpecBit_THDM.cpp
@@ -855,7 +855,7 @@ namespace Gambit
     std::vector<std::vector<complex<double>>> get_qij(const double ba, const double Lam6) {
       const double sba = sin(ba), cba = abs(cos(ba));
       const std::complex<double> i(0.0,1.0);
-      std::vector<std::vector<complex<double>>> q = {{0.0, 0.0, 0.0}, {0.0, sba, sgn(Lam6)*cba}, {0.0, -sgn(Lam6)*cba, sba}, {0.0, 0.0, i}, {0.0, i, 0.0}};
+      std::vector<std::vector<complex<double>>> q = {{0.0, 0.0, 0.0}, {0.0, sba, (double)sgn(Lam6)*cba}, {0.0, -(double)sgn(Lam6)*cba, sba}, {0.0, 0.0, i}, {0.0, i, 0.0}};
       return q;
     }
 
@@ -868,35 +868,35 @@ namespace Gambit
       c += q[j][1]*std::conj(q[k][1])*(q[l][1]).real()*Lam1;
       c += q[j][2]*std::conj(q[k][2])*(q[l][1]).real()*Lam34;
       c += (std::conj(q[j][1])*q[k][2]*q[l][2]*Lam5).real();
-      c += ((2.0*q[j][1] + std::conj(q[j][1]))*std::conj(q[k][1])*q[l][2]*Lam6*sgn(Lam6)).real();
-      c += (std::conj(q[j][2])*q[k][2]*q[l][2]*Lam7*sgn(Lam6)).real();
+      c += ((2.0*q[j][1] + std::conj(q[j][1]))*std::conj(q[k][1])*q[l][2]*Lam6*(double)sgn(Lam6)).real();
+      c += (std::conj(q[j][2])*q[k][2]*q[l][2]*Lam7*(double)sgn(Lam6)).real();
       c *= 0.5*get_v(container);
       return c;
     }
 
     std::complex<double> get_cubic_coupling_hHpHm(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, particle_type k) {
-      return get_v(container)*((q[k][1]).real() * container.higgs_pars.Lambda3 + (q[k][2] * sgn(container.higgs_pars.Lambda6) * container.higgs_pars.Lambda7).real());
+      return get_v(container)*((q[k][1]).real() * container.higgs_pars.Lambda3 + (q[k][2] * (double)sgn(container.higgs_pars.Lambda6) * container.higgs_pars.Lambda7).real());
     }
 
     std::complex<double> get_cubic_coupling_hGpGm(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, particle_type k) {
       double Lambda6 = container.higgs_pars.Lambda6;
-      return get_v(container)*((q[k][1]).real() * container.higgs_pars.Lambda1 + (q[k][2]* sgn(Lambda6) * Lambda6).real());
+      return get_v(container)*((q[k][1]).real() * container.higgs_pars.Lambda1 + (q[k][2]* (double)sgn(Lambda6) * Lambda6).real());
     }
 
     std::complex<double> get_cubic_coupling_hGmHp(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, particle_type k) {
       double Lambda6 = container.higgs_pars.Lambda6;
-      return get_v(container) * 0.5 * sgn(Lambda6) * (std::conj(q[k][2]) * container.higgs_pars.Lambda4 + q[k][2]* container.higgs_pars.Lambda5 + 2.0*(q[k][1]).real()*Lambda6*sgn(Lambda6));
+      return get_v(container) * 0.5 * (double)sgn(Lambda6) * (std::conj(q[k][2]) * container.higgs_pars.Lambda4 + q[k][2]* container.higgs_pars.Lambda5 + 2.0*(q[k][1]).real()*Lambda6*(double)sgn(Lambda6));
     }
 
     std::complex<double> get_cubic_coupling_hhG0(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
       const particle_type k = particles[0], l = particles[1];
       double Lambda6 = container.higgs_pars.Lambda6;
-      return get_v(container) * 0.5 * ( ( q[k][2]*q[l][2]*container.higgs_pars.Lambda5 ).imag() + 2.0*q[k][1]*( q[l][2]*Lambda6*sgn(Lambda6) ).imag() );
+      return get_v(container) * 0.5 * ( ( q[k][2]*q[l][2]*container.higgs_pars.Lambda5 ).imag() + 2.0*q[k][1]*( q[l][2]*Lambda6*(double)sgn(Lambda6) ).imag() );
     }
 
     std::complex<double> get_cubic_coupling_hG0G0(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, particle_type k) {
       double Lambda6 = container.higgs_pars.Lambda6;
-      return get_v(container) * 0.5 * ( q[k][1]*container.higgs_pars.Lambda1 + (q[k][2]*Lambda6*sgn(Lambda6)).real() );
+      return get_v(container) * 0.5 * ( q[k][1]*container.higgs_pars.Lambda1 + (q[k][2]*Lambda6*(double)sgn(Lambda6)).real() );
     }
 
     std::complex<double> get_quartic_coupling_hhhh(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
@@ -909,36 +909,36 @@ namespace Gambit
       c += q[j][2] * q[k][2] * std::conj(q[l][2]) * std::conj(q[m][2]) * Lam2;
       c += 2.0 * q[j][1] * std::conj(q[k][1]) * q[l][2] *std::conj(q[m][2]) * Lam34;
       c += 2.0 * (std::conj(q[j][1]) * std::conj(q[k][1]) * q[l][2] * q[m][2] * Lam5).real();
-      c += 4.0 * (q[j][1] * std::conj(q[k][1]) * std::conj(q[l][1]) * q[m][2] * Lam6 * sgn(Lam6)).real();
-      c += 4.0 * (std::conj(q[j][1]) * q[k][2] * q[l][2] * std::conj(q[m][2]) * Lam7 * sgn(Lam6)).real();
+      c += 4.0 * (q[j][1] * std::conj(q[k][1]) * std::conj(q[l][1]) * q[m][2] * Lam6 * (double)sgn(Lam6)).real();
+      c += 4.0 * (std::conj(q[j][1]) * q[k][2] * q[l][2] * std::conj(q[m][2]) * Lam7 * (double)sgn(Lam6)).real();
       return 0.125*c;
     }
 
     std::complex<double> get_quartic_coupling_hhGpGm(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
       const particle_type j = particles[0], k = particles[1];
       const double Lambda6 = container.higgs_pars.Lambda6;
-      return 0.5 * (q[j][1] * std::conj(q[k][1]) * container.higgs_pars.Lambda1 + q[j][2] * std::conj(q[k][2]) * container.higgs_pars.Lambda3 + 2.0 * (q[j][1] * q[k][2] * Lambda6 * sgn(Lambda6)).real());
+      return 0.5 * (q[j][1] * std::conj(q[k][1]) * container.higgs_pars.Lambda1 + q[j][2] * std::conj(q[k][2]) * container.higgs_pars.Lambda3 + 2.0 * (q[j][1] * q[k][2] * Lambda6 * (double)sgn(Lambda6)).real());
     }
 
     std::complex<double> get_quartic_coupling_hhHpHm(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
       const particle_type j = particles[0], k = particles[1];
-      return 0.5 * (q[j][2] * std::conj(q[k][2]) * container.higgs_pars.Lambda2 + q[j][1] * std::conj(q[k][1]) * container.higgs_pars.Lambda3 + 2.0 * (q[j][1] * q[k][2] * container.higgs_pars.Lambda7 * sgn(container.higgs_pars.Lambda6)).real());
+      return 0.5 * (q[j][2] * std::conj(q[k][2]) * container.higgs_pars.Lambda2 + q[j][1] * std::conj(q[k][1]) * container.higgs_pars.Lambda3 + 2.0 * (q[j][1] * q[k][2] * container.higgs_pars.Lambda7 * (double)sgn(container.higgs_pars.Lambda6)).real());
     }
 
     std::complex<double> get_quartic_coupling_hhGmHp(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
       const particle_type j = particles[0], k = particles[1];
       const double Lam4 = container.higgs_pars.Lambda4, Lam5 = container.higgs_pars.Lambda5, Lam6 = container.higgs_pars.Lambda6, Lam7 = container.higgs_pars.Lambda7;
       std::complex<double> c(0.0,0.0);
-      c += sgn(Lam6) * q[j][1] * std::conj(q[k][2]) * Lam4;
+      c += (double)sgn(Lam6) * q[j][1] * std::conj(q[k][2]) * Lam4;
       c += std::conj(q[j][1]) * q[k][2] * Lam5;
-      c += q[j][1] * std::conj(q[k][1]) * Lam6 * sgn(Lam6);
-      c += q[j][2] * std::conj(q[k][2]) * Lam7 * sgn(Lam6);
+      c += q[j][1] * std::conj(q[k][1]) * Lam6 * (double)sgn(Lam6);
+      c += q[j][2] * std::conj(q[k][2]) * Lam7 * (double)sgn(Lam6);
       return 0.5*c;
     }
 
     std::complex<double> get_quartic_coupling_hG0G0G0(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, particle_type m) {
       const double Lam6 = container.higgs_pars.Lambda6;
-      return 0.5 * (q[m][2] * Lam6 * sgn(Lam6)).imag();
+      return 0.5 * (q[m][2] * Lam6 * (double)sgn(Lam6)).imag();
     }
 
     std::complex<double> get_quartic_coupling_hhG0G0(THDM_spectrum_container& container, std::vector<std::vector<complex<double>>> q, std::vector<particle_type> particles) {
@@ -946,7 +946,7 @@ namespace Gambit
       const double Lam6 = container.higgs_pars.Lambda6;
       std::complex<double> c(0.0,0.0);
       c = (q[l][1] * q[m][1] * container.higgs_pars.Lambda1 + q[l][2] * std::conj(q[m][2]) *( container.higgs_pars.Lambda3 + container.higgs_pars.Lambda4 ));
-      c += -( q[l][2] * q[m][2] * container.higgs_pars.Lambda5 ).real() + 2.0*q[l][1]*( q[m][2]*Lam6*sgn(Lam6) ).real();
+      c += -( q[l][2] * q[m][2] * container.higgs_pars.Lambda5 ).real() + 2.0*q[l][1]*( q[m][2]*Lam6*(double)sgn(Lam6) ).real();
       return 0.25 * c;    
     }
 
@@ -954,8 +954,8 @@ namespace Gambit
       const particle_type k = particles[1], l = particles[1], m = particles[2];
       const double Lam6 = container.higgs_pars.Lambda6;
       std::complex<double> c(0.0,0.0);
-      c = q[k][1] * (q[l][2]*q[m][2]*container.higgs_pars.Lambda5).real() + q[k][1]*q[l][1]*(q[m][2]*Lam6*sgn(Lam6)).real();
-      c += ( q[k][2]*q[l][2]*std::conj(q[m][2])*container.higgs_pars.Lambda7*sgn(Lam6) ).real();
+      c = q[k][1] * (q[l][2]*q[m][2]*container.higgs_pars.Lambda5).real() + q[k][1]*q[l][1]*(q[m][2]*Lam6*(double)sgn(Lam6)).real();
+      c += ( q[k][2]*q[l][2]*std::conj(q[m][2])*container.higgs_pars.Lambda7*(double)sgn(Lam6) ).real();
       return 0.5 * c;    
     }
 
@@ -1014,8 +1014,8 @@ namespace Gambit
 
       int sign = 1;
       for (auto const& each_part : particles) {
-        if(each_part == H0) sign *= -sgn(Lam6);
-        else if (each_part == A0) sign *= sgn(Lam6);
+        if(each_part == H0) sign *= -(double)sgn(Lam6);
+        else if (each_part == A0) sign *= (double)sgn(Lam6);
       }
 
         if (is_neutral(p1) && is_neutral(p2) && is_neutral(p3)) {
@@ -1032,7 +1032,7 @@ namespace Gambit
           else if (particles_match(particles, { p1, Hm, Gp })) c += std::conj(get_cubic_coupling_hGmHp(container, q, p1)); 
         }
 
-      return -i*c*sign;
+      return -i*c*(double)sign;
     }
 
     std::complex<double> get_quartic_coupling(THDM_spectrum_container& container, particle_type p1, particle_type p2, particle_type p3, particle_type p4) {
@@ -1048,8 +1048,8 @@ namespace Gambit
 
       int sign = 1;
       for (auto const& each_part : particles) {
-        if(each_part == H0) sign *= -sgn(Lam6);
-        else if (each_part == A0) sign *= sgn(Lam6);
+        if(each_part == H0) sign *= -(double)sgn(Lam6);
+        else if (each_part == A0) sign *= (double)sgn(Lam6);
       }
 
       for(auto const& particles_perm: get_neutral_particle_permutations(particles)) {
@@ -1078,7 +1078,7 @@ namespace Gambit
         }
       }
       
-      return -i*c*sign;
+      return -i*c*(double)sign;
     }
 
     std::vector<std::complex<double>> get_cubic_couplings(THDM_spectrum_container& container) {
